+++ scripts/organizationSetupCli.ts
@@ -8,6 +8,8 @@ interface CliArguments {
   jupiteroneAccountId: string;
   jupiteroneApiKey: string;
   googleAccessToken: string;
+  skipSystemProjects: boolean;
+  organizationId?: string[] | (string | number)[];
   projectId?: string[] | (string | number)[];
   help?: boolean;
 }
@@ -29,10 +31,10 @@ function redactCliSecret(secret: string) {
   return `${secret.substr(0, 5)}...`;
 }
 
-function projectIdArgToProjectIds(
-  projectId: string | string[] | (string | number)[],
-): string[] | undefined {
-  const projectIds = Array.isArray(projectId) ? projectId : [projectId];
+function convertToArrayOfStrings(
+  data: string | string[] | (string | number)[],
+): string[] {
+  const projectIds = Array.isArray(data) ? data : [data];
   return Array.from(new Set(projectIds.map((p) => p.toString())));
 }
 
@@ -50,11 +52,22 @@ cli
   )
   .option(
     '--google-access-token <googleAccessToken>',
-    '(Required) JupiterOne API Key',
+    '(Required) Google Cloud Access Token',
+  )
+  .option(
+    '--organization-id [organizationId]',
+    '(Optional) Array of organization IDs to collect projects from',
   )
   .option(
     '--project-id [projectId]',
     '(Optional) Array of project IDs to create integration instances with',
+  )
+  .option(
+    '--skip-system-projects [skipSystemProjects]',
+    '(Optional) Skips creation of any projects that have an ID that start with "sys-"',
+    {
+      default: false,
+    },
   )
   .action((options: CliArguments) => {
     assertRequiredCliArg(
@@ -63,6 +76,18 @@ cli
     );
     assertRequiredCliArg('--jupiterone-api-key', options.jupiteroneApiKey);
     assertRequiredCliArg('--google-access-token', options.googleAccessToken);
+
+    if (!options.organizationId && !options.projectId) {
+      throw new Error(
+        'One of the following CLI arguments is required: "--organization-id", "--project-id"',
+      );
+    }
+
+    if (options.organizationId && options.projectId) {
+      throw new Error(
+        'Only one of the following CLI arguments can be provided: "--organization-id", "--project-id"',
+      );
+    }
   });
 
 cli.help();
@@ -73,18 +98,26 @@ const parsed = cli.parse(process.argv, { run: true });
     jupiteroneAccountId,
     jupiteroneApiKey,
     googleAccessToken,
+    organizationId,
     projectId,
     help,
+    skipSystemProjects,
   } = parsed.options as CliArguments;
 
-  const projectIds = projectId && projectIdArgToProjectIds(projectId);
+  const projectIds = projectId && convertToArrayOfStrings(projectId);
+  const organizationIds =
+    organizationId && convertToArrayOfStrings(organizationId);
+  const jupiteroneEnv = process.env.JUPITERONE_ENV;
 
   logger.debug(
     {
       jupiteroneAccountId,
       projectId,
+      organizationIds,
       projectIds,
+      skipSystemProjects,
       jupiteroneApiKey: jupiteroneApiKey && redactCliSecret(jupiteroneApiKey),
+      jupiteroneEnv,
       googleAccessToken:
         googleAccessToken && redactCliSecret(googleAccessToken),
     },
@@ -93,13 +126,21 @@ const parsed = cli.parse(process.argv, { run: true });
 
   if (help) return;
 
-  await setupOrganization({
+  const setupOrganizationResult = await setupOrganization({
     jupiteroneAccountId,
     jupiteroneApiKey,
     googleAccessToken,
     logger,
+    organizationIds,
     projectIds,
+    jupiteroneEnv,
+    skipSystemProjects,
   });
+
+  logger.info(
+    setupOrganizationResult,
+    'Finished setting up all JupiterOne Google Cloud integration instances',
+  );
 })().catch((err) => {
   logger.error({ err }, 'Error running organization setup CLI');
 });